# swagger/v1/swagger.yaml
swagger: "2.0"
info:
  description: "ポートフォリオ用のタスク管理アプリケーション"
  version: "1.0.0"
  title: "Task Management API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: ""
  license:
    name: "MIT"
    url: "https://opensource.org/licenses/MIT"
host: "localhost:3000"
basePath: "/api/v1"
...
schemes:
- "http"
paths:
...
  /auth/sign_out:
    delete:
      tags:
      - "user"
      summary: "log out"
      description: "log out"
      operationId: "devise_token_auth/sessions#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "log out"
        description: "log out"
        required: true
        schema:
          type: object
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: true
        404:
          description: "ログアウト済なのにログアウトした場合のレスポンス"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
                example: false
              errors:
                type: "array"
                items:
                  type: "string"
                  example: "User was not found or was not logged in."
...
  /groups/{group_id}:
    get:
      tags:
      - "group"
      summary: "groupのshow"
      description: "個別groupの参照"
      operationId: "group#show"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "group_id"
        description: "groupのid"
        required: true
      responses:
        201:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "success"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  owner_id:
                    type: "string"
                    example: 3
                  name: 
                    type: "string"
                    example: "test group name"
                  created_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.929Z"
        400:
          description: "groupが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 400
              message:
                type: string
                example: Bad Request
              data:
                type: "array"
                items:
                  type: "string"
                  example: "Name has already been taken"
        401:
          description: "name重複時のレスポンス"
          schema:
            type: "object"
            properties:
              errors:
                type: "array"
                items:
                  type: "string"
                  example: "You need to sign in or sign up before continuing."
      security:
      - access-token: 
        uid:
        client:
    put:
      tags:
      - "group"
      summary: "groupの更新"
      description: "個別groupを指定して情報を更新"
      operationId: "group#update"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "group_id"
        description: "groupのid"
        required: true
      - in: "body"
        name: "group update info"
        description: "json形式で更新情報を作成"
        required: true
        schema:
          type: object
          properties:
            owner_id:
              type: "integer"
              example: 1
            name:
              type: "string"
              example: "test group updated"
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 200
              message:
                type: "string"
                example: "Success Group"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  owner_id:
                    type: "string"
                    example: 3
                  name: 
                    type: "string"
                    example: "test group updated name"
                  created_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.929Z"
        400:
          description: "updateできないときのレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 400
              message:
                type: string
                example: Bad Request
              data:
                type: "array"
                items:
                  type: "string"
                  example: "Name has already been taken"
        401:
          description: "name重複時のレスポンス"
          schema:
            type: "object"
            properties:
              errors:
                type: "array"
                items:
                  type: "string"
                  example: "You need to sign in or sign up before continuing."
        404:
          description: "groupが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: NilClass Not Found
              data:
                type: ""
                example: ""
              notes:
                type: "string"
                notes: "group is not found"
      security:
      - access-token: 
        uid:
        client:
    delete:
      tags:
      - "group"
      summary: "groupの削除"
      description: "個別のgroupの削除"
      operationId: "group#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "group_id"
        description: "groupのid"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 200
              message:
                type: "string"
                example: "Success Group"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  owner_id:
                    type: "string"
                    example: 3
                  name: 
                    type: "string"
                    example: "test group updated name"
                  created_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.929Z"
        400:
          description: "実在しないgroupを指定し、削除失敗しないレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 400
              message:
                type: string
                example: Bad Request
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
...
  /assign/{assign_id}:
    delete:
      tags:
      - "assign"
      summary: "assignを削除する"
      description: "userとgroupの関連付けを削除"
      operationId: "assign#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "assign_id"
        description: "assignのid"
        required: true
      responses:
        201:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 201
              message:
                type: "string"
                example: "Created Job"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  group_id:
                    type: "integer"
                    example: 3
                  title:
                    type: "string"
                    example: "test job title"
                  image:
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  owner_id:
                    type: "integer"
                    example: 2
                  base_date_time:
                    type: "string"
                    example: "2019-05-28T00:00:00.000Z"
                  due_date_time:
                    type: "string"
                    example: "2019-05-29T00:00:00.000Z"
                  frequency:
                    type: "string"
                    example: "weekly"
                  created_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "Assignが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: "NilClass Not Found"
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
...
  /job/{job_id}:
    get:
      tags:
      - "job"
      summary: "jobを参照"
      description: "pathで指定したjobの情報を参照"
      operationId: "job#show"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "job_id"
        description: "jobの参照用"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success Job"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  group_id:
                    type: "integer"
                    example: 3
                  title:
                    type: "string"
                    example: "new test job"
                  image:
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  owner_id:
                    type: "integer"
                    example: 3
                  base_date_time:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  due_date_time:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  frequency:
                    type: "string"
                    example: "weeklys"
                  is_done:
                    type: "boolean"
                    example: false
                  is_approved:
                    type: "boolean"
                    example: false
                  created_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "Assignが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: "NilClass Not Found"
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
    put:
      tags:
      - "job"
      summary: "jobを更新"
      description: "pathで指定したjobの情報を更新"
      operationId: "job#update"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "job_id"
        description: "jobの参照用"
        required: true
      - in: "body"
        name: "job update info"
        description: "jobの参照用"
        required: true
        schema:
          type: object
          properties:
            group_id:
              type: "integer"
              example: 1
            title:
              type: "string"
              example: "test job title update"
            overview:
              type: "string"
              example: "test overview update"
            image:
              type: "string"
              example: "base64encoded image url update"
            owner_id:
              type: "integer"
              example: 1
            base_date_time:
              type: "date"
              example: "2019-06-01T12:01:22.868Z"
            due_date_time:
              type: "date"
              example: "2019-06-02T12:01:22.868Z"
            frequency:
              type: "string"
              example: "weekly"
            is_done:
              type: "boolean"
              example: true
            is_approved:
              type: "boolean"
              example: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success Job"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  group_id:
                    type: "integer"
                    example: 3
                  title:
                    type: "string"
                    example: "new test job"
                  image:
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  owner_id:
                    type: "integer"
                    example: 3
                  base_date_time:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  due_date_time:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  frequency:
                    type: "string"
                    example: "weeklys"
                  is_done:
                    type: "boolean"
                    example: false
                  is_approved:
                    type: "boolean"
                    example: false
                  created_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "Assignが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: "NilClass Not Found"
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
    delete:
      tags:
      - "job"
      summary: "jobを削除"
      description: "pathで指定したjobの情報を削除"
      operationId: "job#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "job_id"
        description: "jobの参照用"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success Job"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  group_id:
                    type: "integer"
                    example: 3
                  title:
                    type: "string"
                    example: "new test job"
                  image:
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  owner_id:
                    type: "integer"
                    example: 3
                  base_date_time:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  due_date_time:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  frequency:
                    type: "string"
                    example: "weeklys"
                  is_done:
                    type: "boolean"
                    example: false
                  is_approved:
                    type: "boolean"
                    example: false
                  created_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "integer"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "Jobが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: "NilClass Not Found"
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
...
  /step/{step_id}:
    get:
      tags:
      - "step"
      summary: "stepを参照"
      description: "pathで指定したstepの情報を参照"
      operationId: "step#show"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "step_id"
        description: "stepの参照用"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success step"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  job_id:
                    type: "integer"
                    example: 3
                  content:
                    type: "string"
                    example: "new test step"
                  assigned_id:
                    type: "integer"
                    example: 3
                  image:
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  due_date:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  is_done:
                    type: "boolean"
                    example: false
                  is_approved:
                    type: "boolean"
                    example: false
                  order:
                    type: "integer"
                    example: 1
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "Assignが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: "NilClass Not Found"
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
    put:
      tags:
      - "step"
      summary: "stepを更新"
      description: "pathで指定したstepの情報を更新"
      operationId: "step#update"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "step_id"
        description: "stepの参照用"
        required: true
      - in: "body"
        name: "step update info"
        description: "stepの参照用"
        required: true
        schema:
          type: object
          properties:
            job_id:
              type: "integer"
              example: 1
            content:
              type: "string"
              example: "test step title"
            assigned_user:
              type: "integer"
              example: 2
            image:
              type: "string"
              example: "base64encoded image url"
            due_date:
              type: "date"
              example: "2019-06-02T12:01:22.868Z"
            is_done:
              type: "boolean"
              example: false
            is_approved:
              type: "boolean"
              example: false
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success step"
              data:
                type: "object"
                properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success step"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  job_id:
                    type: "integer"
                    example: 3
                  content:
                    type: "string"
                    example: "new test step"
                  assigned_id:
                    type: "integer"
                    example: 3
                  image:
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  due_date:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  is_done:
                    type: "boolean"
                    example: false
                  is_approved:
                    type: "boolean"
                    example: false
                  order:
                    type: "integer"
                    example: 1
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "Assignが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: "NilClass Not Found"
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
    delete:
      tags:
      - "step"
      summary: "stepを削除"
      description: "pathで指定したstepの情報を削除"
      operationId: "step#destroy"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "step_id"
        description: "stepの参照用"
        required: true
      responses:
        200:
          description: "成功時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success step"
              data:
                type: "object"
                properties:
              status:
                type: "integer"
                example: 200
              message:
                type: "string"
                example: "Success step"
              data:
                type: "object"
                properties:
                  id:
                    type: "integer"
                    example: 1
                  job_id:
                    type: "integer"
                    example: 3
                  content:
                    type: "string"
                    example: "new test step"
                  assigned_id:
                    type: "integer"
                    example: 3
                  image:
                    type: "string"
                    example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAB4AAAAeCAAAAAAeW/F+AAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAAAmJLR0QA/4ePzL8AAAAHdElNRQfjBRQABBrSkbBKAAAAT0lEQVQoz2M4gxcwjEoPkPSmjau3rNoEFULmQKSPr18/59icM2e27d2J4MClTx05Mn8+SOTUIiQOQve6DRs2rwSat+XAdgRncPh7VBoNAABwbMd4Pq/U3gAAACV0RVh0ZGF0ZTpjcmVhdGUAMjAxOS0wNS0yMFQwOTowNDoyNiswOTowMOwn1y0AAAAldEVYdGRhdGU6bW9kaWZ5ADIwMTktMDUtMjBUMDk6MDQ6MjYrMDk6MDCdem+RAAAAAElFTkSuQmCC"
                  due_date:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  is_done:
                    type: "boolean"
                    example: false
                  is_approved:
                    type: "boolean"
                    example: false
                  order:
                    type: "integer"
                    example: 1
                  created_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.868Z"
                  updated_at:
                    type: "string"
                    example: "2019-06-01T12:01:22.929Z"
        404:
          description: "Stepが実在しない時のレスポンス"
          schema:
            type: "object"
            properties:
              status:
                type: integer
                example: 404
              message:
                type: string
                example: "NilClass Not Found"
              data:
                type: ""
                example: ""
      security:
      - access-token: 
        uid:
        client:
securityDefinitions:
  access-token:
    type: "apiKey"
    name: "access-token"
    in: "header"
    value: "WB4HXdfRAenKUxqgIMvxIw"
  uid:
    type: "apiKey"
    name: "uid"
    in: "header"
    value: "test@test.com"
  client:
    type: "apiKey"
    name: "client"
    in: "header"
    value: "xvUwrt1-gyMsPf1bcqg1tw"